# [REFACTOR] Speed up CertBot Client tests