Improved test speed by 95%!!!

*BEFORE*
Finished in **1.35 seconds** (files took 3.24 seconds to load) 16 examples, 0 failures

*AFTER*
Finished in **0.06339 seconds** (files took 2.86 seconds to load) 16 examples, 0 failures